新取取insert_id方法

+++ mooon/include/mooon/sys/mysql_db.h
@@ -62,7 +62,7 @@ public:
     virtual void reopen() throw (CDBException);
 
     // 如果update的值并没变化返回0，否则返回变修改的行数
-    virtual int update(const char* format, ...) throw (CDBException) __attribute__((format(printf, 2, 3)));
+    virtual uint64_t update(const char* format, ...) throw (CDBException) __attribute__((format(printf, 2, 3)));
     virtual uint64_t get_insert_id() const;
     virtual std::string str() throw ();
