+++ mooon/include/mooon/sys/simple_db.h
@@ -210,7 +210,7 @@ public:
       * 对于MySQL如果update的值并没变化返回0，否则返回变修改的行数
       * 出错则抛出CDBException异常
       */
-    virtual int update(const char* format, ...) throw (CDBException) __attribute__((format(printf, 2, 3))) = 0;
+    virtual uint64_t update(const char* format, ...) throw (CDBException) __attribute__((format(printf, 2, 3))) = 0;
 
     /***
      * 取得insert_id，对MySQL有效
