+++ mooon/include/mooon/sys/sqlite3_db.h
@@ -43,7 +43,7 @@ public:
     virtual void close() throw ();
     virtual void reopen() throw (CDBException);
 
-    virtual int update(const char* format, ...) throw (CDBException) __attribute__((format(printf, 2, 3)));
+    virtual uint64_t update(const char* format, ...) throw (CDBException) __attribute__((format(printf, 2, 3)));
     virtual std::string str() throw ();
 
 private:
